+++ pinyougou-parent/pinyougou-dao/src/main/java/com/pinyougou/mapper/TbSellerMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TbSellerMapper {
-    int countByExample(TbSellerExample example);
+    long countByExample(TbSellerExample example);
 
     int deleteByExample(TbSellerExample example);
