+++ pinyougou-parent/pinyougou-dao/src/main/java/com/pinyougou/mapper/TbSpecificationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TbSpecificationMapper {
-    int countByExample(TbSpecificationExample example);
+    long countByExample(TbSpecificationExample example);
 
     int deleteByExample(TbSpecificationExample example);
